sc2: relocate shadow matrix helper out of common/sc2_utils.h
sc2map links without libshared, but sc2_map.c pulls in sc2_utils.h whose
static sc2_shadow_matrix calls Matrix4_*/Vector3_* from the shared math
library, so make test failed at the sc2map link step. Move the shadow
helper (used only by the renderer and its test) into renderer/sc2/.
